refactor imports to utilize matrix_utils for better organization and functionality
