+++ hardware_accelerators/simulation/__init__.py
@@ -1,6 +1,6 @@
 from .repr_funcs import *
-from .utils import convert_array_dtype, render_waveform
-
+from .matrix_utils import *
+from .utils import render_waveform
 from .systolic import SystolicArraySimulator
 from .accumulators import AccumulatorBankSimulator
 from .buffer import BufferMemorySimulator
